Globally assign 'fetch'
Something changed in one of our libraries (I think we stopped importing
`isomorphic-fetch`) and so now `fetch` is no longer available globally.
However, we still assume that it is available in the `fetch`-related
functions in `controller-utils` and elsewhere. So we still need it to be
around, at least in tests.